@@ -7,27 +7,36 @@
     {
       "name": "Weida Xue",
       "email": "weida@gravitate.co.nz"
+    },
+    {
+      "name": "John Lennard",
+      "email":"john@gravitate.co.nz"
     }
   ],
   "require": {
+    "php": ">=8.1.0",
     "cocur/slugify": "^3.1",
-    "sensio/framework-extra-bundle": "^5.3",
-    "doctrine/dbal": "^2.9",
+    "sensio/framework-extra-bundle": "^6.2.0",
+    "doctrine/dbal": "^3.3.6",
     "symfony/orm-pack": "^1.0",
-    "symfony/twig-bundle": "^4.4",
-    "symfony/form": "^4.4",
-    "symfony/validator": "^4.4",
-    "symfony/translation": "^4.4",
-    "symfony/security-bundle": "^4.4",
-    "symfony/swiftmailer-bundle": "^3.2",
-    "campaignmonitor/createsend-php": "^6.0",
-    "facebook/graph-sdk": "^5.7",
-    "google/apiclient": "^2.2",
-    "phpoffice/phpspreadsheet": "^1.2",
+    "symfony/twig-bundle": "^5.4",
+    "symfony/form": "^5.4",
+    "symfony/validator": "^5.4",
+    "symfony/translation": "^5.4",
+    "symfony/security-bundle": "^5.4",
+    "symfony/mailer": "^5.4",
+    "symfony/routing": "^5.4",
     "bluem/tree": "^3.1",
     "ramsey/uuid": "^4.0",
     "symfony/webpack-encore-bundle": "^1.8",
+    "geoip2/geoip2": "^2.11"
+  },
+  "suggest": {
     "stripe/stripe-php": "^7.75",
+    "phpoffice/phpspreadsheet": "^1.2",
+    "campaignmonitor/createsend-php": "^7.0",
+    "facebook/graph-sdk": "^5.7",
+    "google/apiclient": "^2.2",
     "paypal/paypal-checkout-sdk": "^1.0",
     "geoip2/geoip2": "^2.11",
     "gravitatenz/fta-csp": "^2.1",
@@ -39,6 +48,11 @@
   "autoload": {
     "psr-4": {
       "MillenniumFalcon\\": ""
+    }
+  },
+  "extra": {
+    "branch-alias": {
+      "dev-master": "2.0.x-dev"
     }
   }
 }
