@@ -14,6 +14,9 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpKernel\KernelInterface;
+use Symfony\Component\Mailer\MailerInterface;
+use Symfony\Component\Mime\Email;
+use Symfony\Component\Mime\Message;
 use Twig\Environment;
 
 class FormDescriptorService
@@ -21,32 +24,30 @@ class FormDescriptorService
     /**
      * @var Connection
      */
-    protected $connection;
+    protected Connection $connection;
 
     /**
      * @var KernelInterface
      */
-    protected $kernel;
+    protected KernelInterface $kernel;
 
     /**
      * @var FormFactoryInterface
      */
-    protected $formFactory;
+    protected FormFactoryInterface $formFactory;
 
     /**
      * @var SessionInterface
      */
-    protected $session;
+    protected SessionInterface $session;
 
     /**
      * @var Environment
      */
-    protected $environment;
+    protected Environment $environment;
 
-    /**
-     * @var \Swift_Mailer
-     */
-    protected $mailer;
+    /** @var MailerInterface  */
+    protected MailerInterface $mailer;
 
     /**
      * @var string
@@ -65,7 +66,7 @@ class FormDescriptorService
         FormFactoryInterface $formFactory,
         SessionInterface $session,
         Environment $environment,
-        \Swift_Mailer $mailer
+        MailerInterface $mailer
     ) {
         $this->connection = $connection;
         $this->kernel = $kernel;
@@ -181,7 +182,7 @@ class FormDescriptorService
                     /**
                     if (!$countryInfo) {
                         $request = Request::createFromGlobals();
-                        $countryInfo = UtilsService::ip_info(getenv('TEST_CLIENT_IP') ?: $request->getClientIp());
+                        $countryInfo = UtilsService::ip_info($_ENV['TEST_CLIENT_IP'] ?? $request->getClientIp());
                         $countryInfo = $countryInfo ?: [];
                         $this->session->set(static::COUNTRY_SESSION_KEY, $countryInfo);
                     }
@@ -212,17 +213,20 @@ class FormDescriptorService
                         'submission' => $submission,
                     ));
 
-                    $message = (new \Swift_Message())
-                        ->setSubject("{$formDescriptor->getTitle()} {$submission->getTitle()}")
-                        ->setFrom(array($formDescriptor->getFromAddress()))
-                        ->setTo(array_filter(array_map('trim', explode(',', $formDescriptor->getRecipients()))))
-                        ->setBcc(array(getenv('EMAIL_BCC')))
-                        ->setBody(
-                            $messageBody, 'text/html'
+                    $message = (new Email())
+                        ->subject("{$formDescriptor->getTitle()} {$submission->getTitle()}")
+                        ->from($formDescriptor->getFromAddress())
+                        ->to(...array_filter(array_map('trim', explode(',', $formDescriptor->getRecipients()))))
+                        ->bcc(($_ENV['EMAIL_BCC'] ?? false))
+                        ->html(
+                            $messageBody, 'utf-8'
                         );
+
+
                     if (isset($data['email']) && $data['email'] && filter_var($data['email'], FILTER_VALIDATE_EMAIL)) {
-                        $message->setReplyTo(array($data['email']));
+                        $message->addReplyTo($data['email']);
                     }
+
                     $formDescriptor->sent = $this->mailer->send($message);
 
                     $submission->setEmailStatus($formDescriptor->sent ? 1 : 2);
