@@ -48,7 +48,9 @@ class CheckoutShippingForm extends AbstractType
                 'label' => 'Your email address',
                 'required' => true,
                 'constraints' => [
-                    new Assert\NotBlank(),
+                    new Assert\NotBlank([
+                        'message' => 'Email address is required',
+                    ]),
                     new Assert\Email(),
                 ]
             ])
@@ -60,7 +62,9 @@ class CheckoutShippingForm extends AbstractType
                     'Delivery' => 2,
                 ],
                 'constraints' => [
-                    new Assert\NotBlank(),
+                    new Assert\NotBlank([
+                        'message' => 'Shipping method is required',
+                    ]),
                 ]
             ])
             ->add('pickupFirstName', TextType::class, [
@@ -68,6 +72,7 @@ class CheckoutShippingForm extends AbstractType
                 'required' => true,
                 'constraints' => [
                     new NotBlankIfRequired([
+                        'message' => 'First name is required',
                         'callback' => function ($request) {
                             $data = $request->get($this->getBlockPrefix());
                             return isset($data['isPickup']) && $data['isPickup'] == 1 ? 1 : 0;
@@ -81,6 +86,7 @@ class CheckoutShippingForm extends AbstractType
                 'required' => true,
                 'constraints' => [
                     new NotBlankIfRequired([
+                        'message' => 'Last name is required',
                         'callback' => function ($request) {
                             $data = $request->get($this->getBlockPrefix());
                             return isset($data['isPickup']) && $data['isPickup'] == 1 ? 1 : 0;
@@ -94,6 +100,7 @@ class CheckoutShippingForm extends AbstractType
                 'required' => true,
                 'constraints' => [
                     new NotBlankIfRequired([
+                        'message' => 'Phone number is required',
                         'callback' => function ($request) {
                             $data = $request->get($this->getBlockPrefix());
                             return isset($data['isPickup']) && $data['isPickup'] == 1 ? 1 : 0;
@@ -107,6 +114,7 @@ class CheckoutShippingForm extends AbstractType
                 'required' => true,
                 'constraints' => [
                     new NotBlankIfRequired([
+                        'message' => 'First name is required',
                         'callback' => function ($request) {
                             $data = $request->get($this->getBlockPrefix());
                             return isset($data['isPickup']) && $data['isPickup'] == 2 ? 1 : 0;
@@ -120,6 +128,7 @@ class CheckoutShippingForm extends AbstractType
                 'required' => true,
                 'constraints' => [
                     new NotBlankIfRequired([
+                        'message' => 'Last name is required',
                         'callback' => function ($request) {
                             $data = $request->get($this->getBlockPrefix());
                             return isset($data['isPickup']) && $data['isPickup'] == 2 ? 1 : 0;
@@ -133,6 +142,7 @@ class CheckoutShippingForm extends AbstractType
                 'required' => true,
                 'constraints' => [
                     new NotBlankIfRequired([
+                        'message' => 'Phone number is required',
                         'callback' => function ($request) {
                             $data = $request->get($this->getBlockPrefix());
                             return isset($data['isPickup']) && $data['isPickup'] == 2 ? 1 : 0;
@@ -152,6 +162,7 @@ class CheckoutShippingForm extends AbstractType
                 'required' => true,
                 'constraints' => [
                     new NotBlankIfRequired([
+                        'message' => 'Address is required',
                         'callback' => function ($request) {
                             $data = $request->get($this->getBlockPrefix());
                             return isset($data['isPickup']) && $data['isPickup'] == 2 ? 1 : 0;
@@ -165,6 +176,7 @@ class CheckoutShippingForm extends AbstractType
                 'required' => true,
                 'constraints' => [
                     new NotBlankIfRequired([
+                        'message' => 'City is required',
                         'callback' => function ($request) {
                             $data = $request->get($this->getBlockPrefix());
                             return isset($data['isPickup']) && $data['isPickup'] == 2 ? 1 : 0;
@@ -178,6 +190,7 @@ class CheckoutShippingForm extends AbstractType
                 'required' => true,
                 'constraints' => [
                     new NotBlankIfRequired([
+                        'message' => 'Postcode is required',
                         'callback' => function ($request) {
                             $data = $request->get($this->getBlockPrefix());
                             return isset($data['isPickup']) && $data['isPickup'] == 2 ? 1 : 0;
@@ -187,13 +200,14 @@ class CheckoutShippingForm extends AbstractType
                 ]
             ]);
 
-        if (getenv('SHIPPING_PRICE_MODE') == 1) {
+        if (($_ENV['SHIPPING_PRICE_MODE'] ?? false) == 1) {
             $builder
                 ->add('shippingState', TextType::class, [
                     'label' => 'Region',
                     'required' => true,
                     'constraints' => [
                         new NotBlankIfRequired([
+                            'message' => 'Region is required',
                             'callback' => function ($request) {
                                 $data = $request->get($this->getBlockPrefix());
                                 return isset($data['isPickup']) && $data['isPickup'] == 2 ? 1 : 0;
@@ -208,6 +222,7 @@ class CheckoutShippingForm extends AbstractType
                     'choices' => $countries,
                     'constraints' => [
                         new NotBlankIfRequired([
+                            'message' => 'Country is required',
                             'callback' => function ($request) {
                                 $data = $request->get($this->getBlockPrefix());
                                 return isset($data['isPickup']) && $data['isPickup'] == 2 ? 1 : 0;
@@ -216,7 +231,7 @@ class CheckoutShippingForm extends AbstractType
                         ]),
                     ]
                 ]);
-        } else if (getenv('SHIPPING_PRICE_MODE') == 2) {
+        } else if (($_ENV['SHIPPING_PRICE_MODE'] ?? false) == 2) {
             $builder
                 ->add('shippingCountry', ChoiceType::class, [
                     'label' => 'Country',
@@ -224,6 +239,7 @@ class CheckoutShippingForm extends AbstractType
                     'choices' => $countries,
                     'constraints' => [
                         new NotBlankIfRequired([
+                            'message' => 'Country is required',
                             'callback' => function ($request) {
                                 $data = $request->get($this->getBlockPrefix());
                                 return isset($data['isPickup']) && $data['isPickup'] == 2 ? 1 : 0;
@@ -240,9 +256,21 @@ class CheckoutShippingForm extends AbstractType
                 'required' => true,
                 'constraints' => [
                     new NotBlankIfRequired([
+                        'message' => 'Delivery option is required',
                         'callback' => function ($request) {
                             $data = $request->get($this->getBlockPrefix());
-                            return isset($data['isPickup']) && $data['isPickup'] == 2 ? 1 : 0;
+                            $shippingRequired = isset($data['isPickup']) && $data['isPickup'] == 2 ? 1 : 0;
+
+                            if ($shippingRequired) {
+                                $shippingPriceMode = ($_ENV['SHIPPING_PRICE_MODE'] ?? 1);
+                                if ($shippingPriceMode == 1) {
+                                    return isset($data['shippingState']) && $data['shippingState'] ? 1 : 0;
+                                } else if ($shippingPriceMode == 2) {
+                                    return isset($data['shippingPostcode']) && $data['shippingPostcode'] ? 1 : 0;
+                                }
+                            }
+
+                            return 0;
                         },
                         'request' => $request,
                     ]),
