@@ -78,7 +78,7 @@ trait CmsCoreRestTrait
                 $orm->save(1, [
                     'doNotUpdateModified' => 1,
                 ]);
-                if ($className == '_Model' && ($orm->getIsBuiltIn() != 1 || getenv('ALLOW_CHANGE_BUILTIN') == 1)) {
+                if ($className == '_Model' && ($orm->getIsBuiltIn() != 1 || ($_ENV['ALLOW_CHANGE_BUILTIN'] ?? false) == 1)) {
                     $fullClass::setGenereatedFile($orm, $this->kernel);
                 }
             }
