@@ -9,7 +9,7 @@ class ConstraintRobotValidator extends ConstraintValidator
     public function validate($value, Constraint $constraint)
     {
         $url = 'https://www.google.com/recaptcha/api/siteverify';
-        $myvars = 'response=' . ($_POST['g-recaptcha-response'] ?? "") . '&secret=' . getenv('RECAPTCHA_SECRET_KEY');
+        $myvars = 'response=' . ($_POST['g-recaptcha-response'] ?? "") . '&secret=' . ($_ENV['RECAPTCHA_SECRET_KEY'] ?? false);
 
         $response = file_get_contents("{$url}?{$myvars}");
         $response = json_decode($response);
