@@ -6,6 +6,7 @@ use BlueM\Tree;
 use BlueM\Tree\Serializer\HierarchicalTreeJsonSerializer;
 use Doctrine\DBAL\Connection;
 use MillenniumFalcon\Core\ORM\_Model;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -15,14 +16,15 @@ class AssetService
 {
     const FOLDER_OPEN_MAX_LIMIT = 10;
 
-    /**
-     * AssetService constructor.
-     * @param Connection $connection
-     */
-    public function __construct(Connection $connection)
-    {
-        $this->connection = $connection;
-    }
+    protected static string $uploadPath  = __DIR__ . '/../../../../../uploads/';
+    protected static string $templateFilePath = __DIR__ . '/../../Resources/files/';
+    protected static string $imageCachePath = __DIR__ . '/../../../../../cache/image/';
+
+    public function __construct(
+        protected readonly Connection $connection,
+        protected readonly LoggerInterface $assetLogger,
+    ) {}
+
 
     /**
      * @param $folderRef
@@ -77,6 +79,7 @@ class AssetService
             "order" => 'ASC',
             "orm" => 0,
         ));
+
         foreach ($data as &$itm) {
             if ($itm['id'] == $currentFolderId) {
                 $itm['state'] = [
@@ -95,6 +98,7 @@ class AssetService
 
         $assetRoot = static::getAssetFolderRoot();
         $assetRoot->children = $tree->jsonSerialize();
+
         return $assetRoot;
     }
 
@@ -167,11 +171,13 @@ class AssetService
             'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
             'application/pdf',
             'audio/mpeg',
+            'image/svg+xml',
             'audio/wav',
             'video/x-msvideo',
             'video/mp4',
             'video/mpeg',
         ];
+
 
         if (!in_array($file->getMimeType(), $allowedMimeTypes)) {
             throw new \Exception('Mime type now allowed.');
@@ -193,7 +199,7 @@ class AssetService
      * @param $orm
      * @return JsonResponse
      */
-    static public function processUploadedFileWithAsset(Connection $pdo, UploadedFile $file, $orm)
+    static public function processUploadedFileWithAsset(Connection $pdo, UploadedFile $file, $orm): JsonResponse
     {
         static::removeFile($orm);
         static::removeCaches($pdo, $orm);
@@ -233,7 +239,7 @@ class AssetService
         $fnlFile = $uploadedDir . $orm->getId() . '.' . $ext;
 
         if ($orm->getIsImage() == 1) {
-            $command = getenv('CONVERT_CMD') . ' ' . escapeshellarg($chkFile) . ' -auto-orient ' . escapeshellarg($fnlFile);
+            $command = ($_ENV['CONVERT_CMD'] ?? false) . ' ' . escapeshellarg($chkFile) . ' -auto-orient ' . escapeshellarg($fnlFile);
             static::generateOutput($command);
         } else {
             copy($chkFile, $fnlFile);
@@ -242,7 +248,7 @@ class AssetService
         $orm->setFileLocation($orm->getId() . '.' . $ext);
         $orm->save();
 
-        $SAVE_ASSETS_TO_DB = getenv('SAVE_ASSETS_TO_DB');
+        $SAVE_ASSETS_TO_DB = ($_ENV['SAVE_ASSETS_TO_DB'] ?? false);
         if ($SAVE_ASSETS_TO_DB) {
             $fileLocation = $uploadedDir . $orm->getFileLocation();
             if (file_exists($fileLocation)) {
@@ -308,7 +314,7 @@ class AssetService
      * @param $asset
      * @throws \Exception
      */
-    static public function removeAssetOrms($pdo, $asset)
+    static public function removeAssetOrms($pdo, $asset): void
     {
         $fullClass = ModelService::fullClass($pdo, 'AssetOrm');
         $assetOrms = $fullClass::data($pdo, array(
@@ -323,7 +329,7 @@ class AssetService
     /**
      * @param $asset
      */
-    static public function removeFile($asset)
+    static public function removeFile($asset): void
     {
         $link = static::getUploadPath() . $asset->getFileLocation();
         if (file_exists($link) && is_file($link)) {
@@ -335,9 +341,9 @@ class AssetService
      * @param $pdo
      * @param $asset
      */
-    static public function removeAssetBinary($pdo, $asset)
+    static public function removeAssetBinary($pdo, $asset): void
     {
-        $SAVE_ASSETS_TO_DB = getenv('SAVE_ASSETS_TO_DB');
+        $SAVE_ASSETS_TO_DB = ($_ENV['SAVE_ASSETS_TO_DB'] ?? false);
         if ($SAVE_ASSETS_TO_DB) {
             $assetBinaryFullClass = ModelService::fullClass($pdo, 'AssetBinary');
             $assetBinary = $assetBinaryFullClass::getByField($pdo, 'title', $asset->getId());
@@ -352,7 +358,7 @@ class AssetService
      * @param $asset
      * @throws \Exception
      */
-    static public function removeCaches($pdo, $asset)
+    static public function removeCaches($pdo, $asset): void
     {
         $fullClass = ModelService::fullClass($pdo, 'AssetSize');
         $assetSizes = $fullClass::data($pdo);
@@ -366,7 +372,7 @@ class AssetService
      * @param $asset
      * @throws \Exception
      */
-    static public function removeCachesByAssetSize($pdo, $assetSize)
+    static public function removeCachesByAssetSize($pdo, $assetSize): void
     {
         $fullClass = ModelService::fullClass($pdo, 'Asset');
         $assets = $fullClass::data($pdo, [
@@ -381,7 +387,7 @@ class AssetService
      * @param $asset
      * @param $assetSize
      */
-    static public function removeCache($asset, $assetSize)
+    static public function removeCache($asset, $assetSize): void
     {
         $ext = $asset->getFileExtension();
 
@@ -418,32 +424,38 @@ class AssetService
      * @param $assetSize
      * @return string
      */
-    static public function getCacheKey($asset, $assetSizeCode)
+    static public function getCacheKey($asset, $assetSizeCode): string
     {
         return "{$asset->getCode()}-{$assetSizeCode}";
     }
 
-    /**
-     * @return string
-     */
-    static public function getUploadPath()
+    static public function getUploadPath(): string
     {
-        return __DIR__ . '/../../../../../uploads/';
+        return static::$uploadPath;
     }
 
-    /**
-     * @return string
-     */
-    static public function getImageCachePath()
+    static public function setUploadPath(string $uploadPath): void
     {
-        return __DIR__ . '/../../../../../cache/image/';
+        static::$uploadPath = $uploadPath;
     }
 
-    /**
-     * @return string
-     */
-    static public function getTemplateFilePath()
+    static public function getImageCachePath(): string
     {
-        return __DIR__ . '/../../Resources/files/';
+        return static::$imageCachePath;
+    }
+
+    static public function setImageCachePath(string $imageCachePath): void
+    {
+        static::$imageCachePath = $imageCachePath;
+    }
+
+    static public function getTemplateFilePath(): string
+    {
+        return static::$templateFilePath;
+    }
+
+    static public function setTemplateFilePath(string $templateFilePath): void
+    {
+        static::$templateFilePath = $templateFilePath;
     }
 }
